+++ Coimbra/ITimerService.cs
@@ -15,31 +15,29 @@ namespace Coimbra
         /// </summary>
         /// <param name="timerHandle">It will get reset if not valid anymore.</param>
         /// <returns>True if the timer is still valid and running.</returns>
-        bool IsTimerActive(ref TimerHandle timerHandle);
+        bool IsTimerActive(in TimerHandle timerHandle);
 
         /// <summary>
         /// Starts a new timer.
         /// </summary>
-        /// <param name="timerHandle">It will stop the previously timer if still valid then override it with the new timer handle.</param>
         /// <param name="callback">What should happen when the timer finishes.</param>
         /// <param name="duration">The timer duration.</param>
-        void StartTimer(ref TimerHandle timerHandle, Action callback, float duration);
+        TimerHandle StartTimer(Action callback, float duration);
 
         /// <summary>
         /// Starts a new timer.
         /// </summary>
-        /// <param name="timerHandle">It will stop the previously timer if still valid then override it with the new timer handle.</param>
         /// <param name="callback">What should happen each time the timer triggers.</param>
         /// <param name="delay">The delay to trigger the first time.</param>
         /// <param name="rate">The interval between each trigger.</param>
         /// <param name="loops">The amount of times it should trigger. Infinite if 0 or negative.</param>
-        void StartTimer(ref TimerHandle timerHandle, Action callback, float delay, float rate, int loops = 0);
+        TimerHandle StartTimer(Action callback, float delay, float rate, int loops = 0);
 
         /// <summary>
         /// Stops an existing timer if still valid.
         /// </summary>
         /// <param name="timerHandle">Ignored if not valid anymore.</param>
-        void StopTimer(ref TimerHandle timerHandle);
+        void StopTimer(in TimerHandle timerHandle);
 
         /// <summary>
         /// Stops all existing timers.