+++ Coimbra/TimerHandle.cs
@@ -5,13 +5,10 @@ namespace Coimbra
 {
     [Preserve]
     [Serializable]
-    public struct TimerHandle : IEquatable<TimerHandle>
+    public readonly struct TimerHandle : IEquatable<TimerHandle>
     {
-        public Guid Guid { get; private set; }
+        public readonly Guid Guid;
 
-        /// <summary>
-        /// Expected to be called from within a <see cref="ITimerService"/>.
-        /// </summary>
         public TimerHandle(Guid guid)
         {
             Guid = guid;
@@ -42,22 +39,6 @@ namespace Coimbra
         public override int GetHashCode()
         {
             return Guid.GetHashCode();
-        }
-
-        /// <summary>
-        /// Expected to be called from within a <see cref="ITimerService"/>.
-        /// </summary>
-        public void Initialize(Guid guid)
-        {
-            Guid = guid;
-        }
-
-        /// <summary>
-        /// Expected to be called from within a <see cref="ITimerService"/>.
-        /// </summary>
-        public void Invalidate()
-        {
-            Guid = Guid.Empty;
         }
     }
 }